+++ WorldServer/Entities/CGuild.h
@@ -157,7 +157,7 @@ public:
   void SendMsg_ManageGuildCommitteeResult(char bAppoint, char *pwszCommitteeName);
   void SendMsg_MasterElectPossible(char bPossible);
   void Complete_DB_Update_Committee(char *pData);
-  void CompleteSelectMasterLastConn(unsigned int dwLastConnTime);
+  void CompleteSelectMasterLastConn(unsigned long long dwLastConnTime);
   unsigned int GetMemberNum();
   void Release();
